// src/app/models/new-user.ts
export class NewUser {
    evenement: string;
    onderdeel: string;
    voornaam: string;
    tussenvoegsel?: string;
    achternaam: string;
    adres: string;
    postcode: string;
    woonplaats: string;
    land: string;
    geboortedatum: string;
    geslacht: string;
    telefoonnummer: number;
    email: string;
    nieuwsbrief: boolean;
    voorwaarden: boolean;
}